+++ linux/linux_native.h
@@ -22,11 +22,11 @@ public:
 	void UnregisterHotKey(uint32_t key, uint32_t modifier) override;
 	void SendKeysDown(uint32_t* keys, size_t count) override;
 	void SendKeysUp(uint32_t* keys, size_t count) override;
-	void SetMousePos(double x, double y) override;
+	void SetMousePos(int x, int y) override;
 	bool SetFocusOnProcess(const std::string& process_name) override;
 	void CursorHide(bool hide) override;
 	void Update() override;
-	void GetMousePos(double* x_ret, double* y_ret) override;
+	void GetMousePos(int* x_ret, int* y_ret) override;
 
 private:
 	struct ScanCodeInfo
