+++ native.h
@@ -28,8 +28,8 @@ public:
 	virtual void UnregisterHotKey(uint32_t key, uint32_t modifier) = 0;
 	virtual void SendKeysDown(uint32_t* keys, size_t count) = 0;
 	virtual void SendKeysUp(uint32_t* keys, size_t count) = 0;
-	virtual void SetMousePos(double x, double y) = 0;
-	virtual void GetMousePos(double* x_ret, double* y_ret) = 0;
+	virtual void SetMousePos(int x, int y) = 0;
+	virtual void GetMousePos(int* x_ret, int* y_ret) = 0;
 	/* optional */
 	virtual bool SetFocusOnProcess(const std::string& process_name) = 0;
 	virtual void CursorHide(bool hide) = 0;
